backend/pyproject.toml
[project]
name = "react-rag-backend"
version = "0.1.0"
description = "FastAPI backend for React RAG PDF app"
authors = [
    {name = "kamya"}
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi (==0.111.0)",
    "uvicorn[standard] (==0.29.0)",
    "python-multipart (==0.0.9)",
    "pydantic (==2.7.1)",
    "groq (>=0.9.0)",
    "httpx (<0.28.0)",
    "pymupdf (==1.24.3)",
    "sentence-transformers (==2.7.0)",
    "faiss-cpu (==1.8.0)",
    "numpy (==1.26.4)",
    "python-dotenv (==1.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
